pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "numpy>=2.2.4",
    "opentelemetry-api>=1.31.1",
    "opentelemetry-exporter-otlp>=1.31.1",
    "opentelemetry-instrumentation-flask>=0.52b1",
    "opentelemetry-sdk>=1.31.1",
    "psycopg2-binary>=2.9.10",
    "scapy>=2.6.1",
    "speechrecognition>=3.14.2",
]
